refactor size styles and themeContextColor
